Wait for the body if the init message
The header and the body are sent in two steps. We might have to wait for
the body to be received. It fix a crash occuring in the checksum
computation when the body is empty.

The wait occurs synchronously, we might want to register a signal and do
all the processing in slots instead. However, we will have to track the
parsing state, which increase a bit the complexity.

This patch also changes the raw delete by deleteLater() which is a more Qtish
way of managing the object lifespan.

The patch also checks that the init message length is at least the
checksum size and also smaller than a maximum. If the header says that
the init message is greater than 128, it is certainly a bug.

Signed-off-by: Francis Giraldeau <francis.giraldeau@nrc-cnrc.gc.ca>